move CommentAction to its own file
